debundle/lowering: extract plans
BAZEL_TEST_INVOCATIONS=none: file move.
