+++ devinfra/js/debundle/lowering/mod.rs
@@ -40,6 +40,7 @@ mod imports_cross;
 mod imports_runtime;
 mod naturalize;
 mod plan_references;
+mod plans;
 mod rewrite_runtime;
 mod runtime_imports;
 mod visitors;
@@ -66,6 +67,10 @@ use naturalize::naturalize_module_body;
 use plan_references::{
     ArtifactSourceImportResolutionCache, EntryExport, ModuleReferenceNeeds, RuntimeImportLookup,
     collect_imported_reexports_by_module, plan_module_reference_needs,
+};
+use plans::{
+    LogicalRequest, MemberRequest, ModulePlan, known_effect_from_member_effect,
+    logical_requests_for_chunk, synthesize_mini_factor_plans,
 };
 use rewrite_runtime::rewrite_runtime_sources_for_target;
 use runtime_imports::{
@@ -184,75 +189,6 @@ pub struct MaterializeLogicalModulesOptions {
     pub report_out_dir: Option<PathBuf>,
     pub report_summary_path: Option<PathBuf>,
     pub target_dir: String,
-}
-
-#[derive(Debug, Clone)]
-struct LogicalRequest {
-    id: String,
-    target_path: String,
-    residual: bool,
-    members: Vec<MemberRequest>,
-    /// Verbatim source of each anonymous-statement member the spec
-    /// asked to co-move into this module. Resolved later (after AST
-    /// analysis) into [`ModulePlan::anonymous_statement_ordinals`].
-    anonymous_match_sources: Vec<String>,
-}
-
-#[derive(Debug, Clone)]
-struct MemberRequest {
-    binding: String,
-    export_name: String,
-    /// When `true`, the member's source is an import specifier in the
-    /// source chunk (not a top-level decl). The materializer looks up
-    /// the import statement by `binding` in the chunk body and rewrites
-    /// it to a re-import in the destination module.
-    is_import_specifier: bool,
-    /// Spec-level purity annotation. `Pure` asserts that calls to the
-    /// bound function have no observable side effects — the validator
-    /// trusts the annotation and drops S edges for `<binding>(...)`
-    /// call sites. `Default` means "not annotated, fall back to
-    /// inferred classification". An author-trust contract; see
-    /// AGENTS.md "Declared purity" and DESIGN.md A9.
-    purity: MemberPurity,
-    /// Spec-level local-effect annotation. `TypescriptDecorateHelper`
-    /// asserts that recognized calls to the bound helper mutate only
-    /// their target class/prototype, so the analyzer can model a local
-    /// effect edge instead of a global side-effect-order edge.
-    effect: MemberEffect,
-    /// Property names on the bound value whose member calls
-    /// (`<binding>.<prop>(args)` / `<binding>?.<prop>(args)`) the author
-    /// asserts have no observable side effects beyond evaluating their
-    /// arguments. Same author-trust contract as `purity: pure` — see
-    /// AGENTS.md "Declared purity". Empty when the spec doesn't carry a
-    /// `pure_members` entry for this member.
-    pure_members: Vec<String>,
-}
-
-#[derive(Debug, Clone)]
-struct ModulePlan {
-    id: String,
-    target_file: String,
-    /// Logical module path the spec asked for (e.g. `"ai/mcp/foo"`).
-    /// Distinct from `target_file`, which is the chunk-relative
-    /// emitted file path (e.g. `"modules/foo.js"`).
-    target_path: String,
-    explicit: bool,
-    /// Local-name → public-export-name for every owned binding this
-    /// plan claims (i.e. members whose `selector.binding.kind` is
-    /// _not_ `ImportSpecifier`). ImportSpecifier-bound members live
-    /// in `Schedule.bindings` as `BindingKind::Imported` and their
-    /// emit is driven from there. Iteration order is undefined;
-    /// emit / report sites sort by local name before consuming so
-    /// the emitted source and JSON shapes stay deterministic.
-    bindings: HashMap<String, String>,
-    /// Source-chunk statement ordinals of anonymous-statement members
-    /// claimed by this module. These owners have empty
-    /// `declared_bindings`, so they can't be addressed by name —
-    /// the spec resolves them by AST shape (see
-    /// [`spec::LogicalModule::anonymous_statements`]). The
-    /// materializer routes each such statement into this module's
-    /// body in source order, alongside the named members.
-    anonymous_statement_ordinals: Vec<usize>,
 }
 
 pub fn materialize_logical_modules(
@@ -1804,84 +1740,6 @@ fn lower_chunk(inputs: LowerChunkInputs<'_>) -> Result<LoweredChunk> {
     })
 }
 
-fn logical_requests_for_chunk(
-    chunk_logical_modules: Option<&BTreeMap<String, LogicalModule>>,
-    chunk_unassigned_mode: &UnassignedMode,
-    chunk_renames_present: bool,
-    chunk_id: &str,
-    target_dir: &str,
-) -> Result<Vec<LogicalRequest>> {
-    let mut requests = Vec::new();
-    let catchall_target = chunk_unassigned_mode
-        .catchall_file_target()
-        .map(str::to_string);
-    let mut explicit_module_at_catchall = false;
-    if let Some(by_target_path) = chunk_logical_modules {
-        for (target_path, module) in by_target_path {
-            let id = format!("{chunk_id}::{target_path}");
-            let members = build_members(&module.members);
-            reject_duplicate_export_names("logical_module", &id, &members)?;
-            reject_duplicate_member_bindings("logical_module", &id, &members)?;
-            let anonymous_match_sources = module
-                .anonymous_statements
-                .iter()
-                .map(|stmt| stmt.match_source.clone())
-                .collect();
-            if catchall_target.as_deref() == Some(target_path.as_str()) {
-                explicit_module_at_catchall = true;
-            }
-            requests.push(LogicalRequest {
-                id,
-                target_path: target_path.clone(),
-                residual: false,
-                members,
-                anonymous_match_sources,
-            });
-        }
-    }
-    // Synthesize a memberless catchall-file request when the chunk's
-    // `unassigned_mode` is `CatchallFile` and no explicit logical
-    // module already claims the catchall target. When an explicit
-    // module *is* at the catchall target, the residual sweep in
-    // `materialize_logical_chunk` will append unclaimed bindings to
-    // that explicit plan instead.
-    if let Some(target_path) = catchall_target
-        && !explicit_module_at_catchall
-    {
-        requests.push(LogicalRequest {
-            id: format!("{chunk_id}::residual"),
-            target_path,
-            residual: true,
-            members: Vec::new(),
-            anonymous_match_sources: Vec::new(),
-        });
-    }
-    // Fallback: when the spec is silent about this chunk (no
-    // `logical_modules`, default `InlineInEntry` mode, no
-    // `chunk_renames`), inject a memberless residual so the
-    // materializer has at least one module to point unowned decls
-    // at. Skipped when the spec has any `chunk_renames` for the
-    // chunk — that signals the spec wants bindings to stay in
-    // `ResidualEntry`-land (no `Logical(R)` module, no separate
-    // residual file emitted), with renames applied in-place by the
-    // lowerer. Skipped when `MiniFactors` is active — the
-    // synthesizer takes care of placing unclaimed code into
-    // mini-factor modules.
-    if requests.is_empty()
-        && !chunk_renames_present
-        && !matches!(chunk_unassigned_mode, UnassignedMode::MiniFactors)
-    {
-        requests.push(LogicalRequest {
-            id: format!("{chunk_id}::residual"),
-            target_path: join_module_path(&[target_dir, "unhandled"]),
-            residual: true,
-            members: Vec::new(),
-            anonymous_match_sources: Vec::new(),
-        });
-    }
-    Ok(requests)
-}
-
 /// Collect a `ChunkRenames` block into a `binding_name -> export_name`
 /// map. Bindings that appear more than once across `members` fail
 /// fast — silently last-write-wins on a binding rename is the same
@@ -1961,166 +1819,6 @@ fn body_index_for_statement_ordinal(body: &[ModuleItem], stmt_ordinal: usize) ->
 /// only holds whatever truly couldn't be peeled (typically nothing
 /// for clean chunks).
 ///
-/// The synthesized plan's `target_path` is deterministic
-/// (`__auto/mini/{idx:04}`) and indexed by the unit's position in
-/// the iteration order of unclaimed units, sorted by the smallest
-/// member `OwnerId` so the names are stable run-to-run.
-#[allow(clippy::too_many_arguments)]
-fn synthesize_mini_factor_plans(
-    precomputed: &OwnerGraphAndUnits,
-    body: &[ModuleItem],
-    residual_plan_index: Option<usize>,
-    module_plans: &mut Vec<ModulePlan>,
-    binding_assignment: &mut BTreeMap<String, usize>,
-    bindings_catalogue: &mut HashMap<Id, BindingKind>,
-    anonymous_ordinal_assignment: &mut BTreeMap<usize, usize>,
-    chunk_top_level_mark: swc_common::Mark,
-    target_dir: &str,
-) -> Result<()> {
-    let owner_graph = &precomputed.owner_graph;
-    let atomic_units = &precomputed.atomic_units;
-    let mut owner_declared_names: HashMap<OwnerId, Vec<BindingName>> = HashMap::new();
-    let mut owner_statement_ordinal: HashMap<OwnerId, usize> = HashMap::new();
-    for node in owner_graph.iter_nodes() {
-        let names: Vec<BindingName> = node
-            .declared
-            .iter()
-            .filter_map(|bid| owner_graph.binding_table.name(*bid).cloned())
-            .collect();
-        owner_declared_names.insert(node.id, names);
-        owner_statement_ordinal.insert(node.id, node.statement_ordinal.0);
-    }
-
-    // A unit member counts as unclaimed iff every declared binding is
-    // either absent from `binding_assignment` or assigned to the
-    // residual plan (if any); anonymous owners must similarly be
-    // unassigned or routed via residual. If any member is claimed by
-    // an explicit (non-residual) plan, the spec author already named
-    // the unit's destination — leave the existing claim intact (and
-    // let downstream validation flag an atomic-unit conflict if the
-    // claims disagree).
-    let is_owner_unclaimed = |owner: OwnerId| -> bool {
-        let names = owner_declared_names
-            .get(&owner)
-            .map(Vec::as_slice)
-            .unwrap_or(&[]);
-        for name in names {
-            match binding_assignment.get(name).copied() {
-                None => continue,
-                Some(idx) if Some(idx) == residual_plan_index => continue,
-                Some(_) => return false,
-            }
-        }
-        if names.is_empty() {
-            let Some(stmt_ord) = owner_statement_ordinal.get(&owner).copied() else {
-                return true;
-            };
-            let Some(body_idx) = body_index_for_statement_ordinal(body, stmt_ord) else {
-                return true;
-            };
-            match anonymous_ordinal_assignment.get(&body_idx).copied() {
-                None => return true,
-                Some(idx) if Some(idx) == residual_plan_index => return true,
-                Some(_) => return false,
-            }
-        }
-        true
-    };
-
-    let mut unclaimed_units: Vec<&BTreeSet<OwnerId>> = atomic_units
-        .iter()
-        .filter(|unit| unit.members.iter().copied().all(is_owner_unclaimed))
-        .map(|unit| &unit.members)
-        .collect();
-    // Stable iteration order: smallest OwnerId first.
-    unclaimed_units.sort_by_key(|members| members.iter().next().copied());
-
-    for (idx, members) in unclaimed_units.into_iter().enumerate() {
-        let synthetic_idx = module_plans.len();
-        let synthetic_module_id = ModuleId(LogicalModuleIndex(synthetic_idx));
-        let target_path = format!("__auto/mini/{idx:04}");
-        let target_file = target_file_for_request(target_dir, &target_path)?;
-        let mut bindings = HashMap::<String, String>::new();
-        let mut anonymous_statement_ordinals = Vec::<usize>::new();
-        for owner in members {
-            let names = owner_declared_names
-                .get(owner)
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            if names.is_empty() {
-                let Some(stmt_ord) = owner_statement_ordinal.get(owner).copied() else {
-                    continue;
-                };
-                let Some(body_idx) = body_index_for_statement_ordinal(body, stmt_ord) else {
-                    continue;
-                };
-                anonymous_ordinal_assignment.insert(body_idx, synthetic_idx);
-                anonymous_statement_ordinals.push(body_idx);
-                continue;
-            }
-            for name in names {
-                bindings.insert(name.clone(), name.clone());
-                // Move the binding out of the residual plan (if it was
-                // staged there by the sweep above) into the synthesized
-                // plan. The residual plan's bindings/anonymous-ordinal
-                // maps are pruned so it doesn't double-claim members.
-                if let Some(prev) = binding_assignment.get(name).copied() {
-                    if Some(prev) == residual_plan_index {
-                        if let Some(residual_idx) = residual_plan_index {
-                            module_plans[residual_idx].bindings.remove(name);
-                        }
-                    }
-                }
-                binding_assignment.insert(name.clone(), synthetic_idx);
-                bindings_catalogue.insert(
-                    top_level_id(name.as_str(), chunk_top_level_mark),
-                    BindingKind::Owned {
-                        owner: synthetic_module_id,
-                    },
-                );
-            }
-        }
-        anonymous_statement_ordinals.sort_unstable();
-        module_plans.push(ModulePlan {
-            id: target_path.clone(),
-            target_file,
-            target_path,
-            explicit: false,
-            bindings,
-            anonymous_statement_ordinals,
-        });
-    }
-    Ok(())
-}
-
-fn build_members(members: &[spec::Member]) -> Vec<MemberRequest> {
-    members
-        .iter()
-        .map(|m| {
-            let binding = m.selector.binding.name.clone();
-            let export_name = m.name.clone().unwrap_or_else(|| binding.clone());
-            MemberRequest {
-                is_import_specifier: matches!(
-                    m.selector.binding.kind,
-                    Some(BindingSourceKind::ImportSpecifier)
-                ),
-                binding,
-                export_name,
-                purity: m.purity,
-                effect: m.effect,
-                pure_members: m.pure_members.clone(),
-            }
-        })
-        .collect()
-}
-
-fn known_effect_from_member_effect(effect: MemberEffect) -> Option<KnownEffect> {
-    match effect {
-        MemberEffect::Default => None,
-        MemberEffect::TypescriptDecorateHelper => Some(KnownEffect::TypescriptDecorateHelper),
-    }
-}
-
 fn is_identifier_like(name: &str) -> bool {
     let mut chars = name.chars();
     let Some(first) = chars.next() else {
